requirements.txt
Flask-Inputs==0.3.0
Flask==1.0.2
google-cloud-bigquery==2.16.0
google-cloud-core==1.6.0
google-cloud-pubsub==2.4.1
jsonschema==3.0.1
pytest==4.6.2